Refactor: notificationMapper to notificationRepository
