+++ src/main/java/com/rollingpaper/ggeujeogggeujeog/notification/application/NotificationFeedService.java
@@ -6,9 +6,9 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.rollingpaper.ggeujeogggeujeog.notification.domain.Notification;
+import com.rollingpaper.ggeujeogggeujeog.notification.domain.NotificationRepository;
 import com.rollingpaper.ggeujeogggeujeog.notification.domain.NotificationStatus;
 import com.rollingpaper.ggeujeogggeujeog.notification.exception.NoSuchNotification;
-import com.rollingpaper.ggeujeogggeujeog.notification.infrastructure.NotificationMapper;
 import com.rollingpaper.ggeujeogggeujeog.notification.presentation.dto.NotificationsDto;
 
 import lombok.RequiredArgsConstructor;
@@ -17,11 +17,11 @@ import lombok.RequiredArgsConstructor;
 @Service
 public class NotificationFeedService {
 
-	private final NotificationMapper notificationMapper;
+	private final NotificationRepository notificationRepository;
 
 	@Transactional(readOnly = true)
 	public NotificationsDto getUserNotifications(Long userId) {
-		List<Notification> notificationList = notificationMapper.findAllByUserId(userId);
+		List<Notification> notificationList = notificationRepository.findAllByUserId(userId);
 		return NotificationsDto.from(notificationList);
 	}
 
@@ -29,12 +29,12 @@ public class NotificationFeedService {
 	public void deleteNotification(Long notificationId) {
 		Notification notification = findById(notificationId);
 		notification.updateStatus(NotificationStatus.DELETED);
-		notificationMapper.update(notification);
+		notificationRepository.update(notification);
 	}
 
 	@Transactional
 	public Notification findById(Long notificationId) {
-		return notificationMapper.findNotificationById(notificationId)
+		return notificationRepository.findNotificationById(notificationId)
 			.orElseThrow(NoSuchNotification::new);
 	}
 }
